feat: move email campaigns dashboard to summary
