+++ Doppler.ReportingApi.Test/Controllers/SummaryControllerTest.cs
@@ -8,25 +8,22 @@ using Moq.Dapper;
 using System;
 using System.Collections.Generic;
 using System.Data.Common;
+using System.Data;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
-using System.Text;
 using System.Threading.Tasks;
 using Xunit;
-using Xunit.Abstractions;
 
 namespace Doppler.ReportingApi.Controllers
 {
     public class SummaryControllerTest : IClassFixture<WebApplicationFactory<Startup>>
     {
         private readonly WebApplicationFactory<Startup> _factory;
-        private readonly ITestOutputHelper _output;
 
-        public SummaryControllerTest(WebApplicationFactory<Startup> factory, ITestOutputHelper output)
+        public SummaryControllerTest(WebApplicationFactory<Startup> factory)
         {
             _factory = factory;
-            _output = output;
         }
 
         [Fact]
@@ -136,6 +133,66 @@ namespace Doppler.ReportingApi.Controllers
             // Assert
             Assert.Equal(HttpStatusCode.OK, response.StatusCode);
             Assert.NotNull(content);
+        }
+
+        [Fact]
+        public async Task Get_email_campaigns_dashboard_should_return_valid_response()
+        {
+            // Arrange
+            var userName = "test1@test.com";
+            var token = TestJwtTokenFactory.ValidAccount123Test1;
+            var mockConnection = new Mock<DbConnection>();
+
+            mockConnection
+                .SetupDapperAsync(c => c.QueryAsync<EmailCampaignDashboardItem>(It.IsAny<string>(), It.IsAny<object>(), null, null, It.IsAny<CommandType?>()))
+                .ReturnsAsync(new List<EmailCampaignDashboardItem>());
+
+            var client = _factory.WithWebHostBuilder(builder =>
+            {
+                builder.ConfigureTestServices(services =>
+                {
+                    services.SetupConnectionFactory(mockConnection.Object);
+                });
+            }).CreateClient(new WebApplicationFactoryClientOptions());
+
+            // Act
+            var response = await client.SendAsync(new HttpRequestMessage(
+                HttpMethod.Get,
+                $"/{userName}/dashboard/email-campaigns?startDate=2025-06-10T00:00:00Z&endDate=2025-06-11T00:00:00Z&campaignType=regular")
+            {
+                Headers = { { "Authorization", $"Bearer {token}" } }
+            });
+
+            // Assert
+            Assert.Equal(HttpStatusCode.OK, response.StatusCode);
+        }
+
+        [Fact]
+        public async Task Get_email_campaigns_dashboard_should_return_bad_request_when_dates_are_missing()
+        {
+            // Arrange
+            var userName = "test1@test.com";
+            var token = TestJwtTokenFactory.ValidAccount123Test1;
+            var mockConnection = new Mock<DbConnection>();
+
+            var client = _factory.WithWebHostBuilder(builder =>
+            {
+                builder.ConfigureTestServices(services =>
+                {
+                    services.SetupConnectionFactory(mockConnection.Object);
+                });
+            }).CreateClient(new WebApplicationFactoryClientOptions());
+
+            // Act
+            var response = await client.SendAsync(new HttpRequestMessage(
+                HttpMethod.Get,
+                $"/{userName}/dashboard/email-campaigns?startDate=2025-06-10T00:00:00Z&campaignType=regular")
+            {
+                Headers = { { "Authorization", $"Bearer {token}" } }
+            });
+
+            // Assert
+            Assert.Equal(HttpStatusCode.BadRequest, response.StatusCode);
         }
     }
 }
