+++ Doppler.ReportingApi/Infrastructure/ISummaryRepository.cs
@@ -10,6 +10,7 @@ namespace Doppler.ReportingApi.Infrastructure
         Task<CampaignsSummary> GetCampaignsSummaryByUserAsync(string userName, DateTime startDate, DateTime endDate);
         Task<SubscribersSummary> GetSubscribersSummaryByUserAsync(string userName, DateTime startDate, DateTime endDate);
         Task<IEnumerable<SubscriberStatusStat>> GetSubscribersDashboardByUserAsync(string userName, DateTime startDate, DateTime endDate);
+        Task<IEnumerable<EmailCampaignDashboardItem>> GetEmailCampaignsAsync(string accountName, DateTime startDate, DateTime endDate, string campaignType);
         Task<SystemUsageSummary> GetSystemUsageAsync(string accountName);
     }
 }
