+++ Doppler.ReportingApi/Infrastructure/InfrastructureServiceCollectionExtensions.cs
@@ -13,7 +13,6 @@ namespace Microsoft.Extensions.DependencyInjection
             services.AddScoped<IDatabaseConnectionFactory, DatabaseConnectionFactory>();
             services.AddScoped<ISummaryRepository, SummaryRepository>();
             services.AddScoped<ICampaignRepository, CampaignRepository>();
-            services.AddScoped<IDashboardRepository, DashboardRepository>();
             services.AddTransient<JwtSecurityTokenHandler>();
             services.AddScoped<IJwtTokenGenerator, JwtTokenGenerator>();
             services.AddScoped<ISuperUserTokenService, SuperUserTokenService>();
